--- include/aquarius_protocol.hpp
@@ -1,18 +1,0 @@
-#pragma once
-#include <aquarius_protocol/basic_http_header.hpp>
-#include <aquarius_protocol/basic_http_message.hpp>
-#include <aquarius_protocol/basic_message.hpp>
-#include <aquarius_protocol/basic_tcp_message.hpp>
-#include <aquarius_protocol/binary.hpp>
-#include <aquarius_protocol/concepts.hpp>
-#include <aquarius_protocol/custom_tcp_header.hpp>
-#include <aquarius_protocol/http_filed_type.hpp>
-#include <aquarius_protocol/http_method.hpp>
-#include <aquarius_protocol/http_request.hpp>
-#include <aquarius_protocol/http_response.hpp>
-#include <aquarius_protocol/http_status.hpp>
-#include <aquarius_protocol/json.hpp>
-#include <aquarius_protocol/layer_tcp_header.hpp>
-#include <aquarius_protocol/reflection.hpp>
-#include <aquarius_protocol/tcp_request.hpp>
-#include <aquarius_protocol/tcp_response.hpp>
