--- include/aquarius_protocol/http_response.hpp
@@ -1,8 +1,0 @@
-#pragma once
-#include <aquarius_protocol/basic_http_message.hpp>
-
-namespace aquarius
-{
-	template<typename Body>
-	using http_response = basic_http_message<false, Body>;
-}
