--- include/aquarius_protocol/reflection.hpp
@@ -1,7 +1,0 @@
-#pragma once
-#include <boost/pfr.hpp>
-
-namespace aquarius
-{
-	 namespace reflect = boost::pfr;
-}
