+++ include/virgo/detail/reflection.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <boost/pfr.hpp>
+
+namespace virgo
+{
+	namespace reflect = boost::pfr;
+}
