+++ include/virgo/http.hpp
@@ -1,0 +1,3 @@
+#pragma once
+#include <virgo/http/request.hpp>
+#include <virgo/http/response.hpp>
