+++ include/virgo/http/request.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include <virgo/detail/config.hpp>
+#include <virgo/http/detail/basic_message.hpp>
+
+namespace virgo
+{
+	namespace http
+	{
+		template <typename Body>
+		using request = detail::basic_message<true, Body>;
+	}
+} // namespace virgo
