+++ include/virgo/tcp.hpp
@@ -1,0 +1,3 @@
+#pragma once
+#include <virgo/tcp/request.hpp>
+#include <virgo/tcp/response.hpp>
